# PERF: delta_to_nanoseconds