--- a/pandas/_libs/tslibs/timedeltas.pxd
+++ b/pandas/_libs/tslibs/timedeltas.pxd
@@ -6,10 +6,11 @@ from .np_datetime cimport NPY_DATETIMEUNIT
 
 # Exposed for tslib, not intended for outside use.
 cpdef int64_t delta_to_nanoseconds(
-    delta, NPY_DATETIMEUNIT reso=*, bint round_ok=*, bint allow_year_month=*
+    delta, NPY_DATETIMEUNIT reso=*, bint round_ok=*
 ) except? -1
 cdef convert_to_timedelta64(object ts, str unit)
 cdef bint is_any_td_scalar(object obj)
+cdef object ensure_td64ns(object ts)
 
 
 cdef class _Timedelta(timedelta):