--- a/pandas/tests/scalar/timestamp/test_arithmetic.py
+++ b/pandas/tests/scalar/timestamp/test_arithmetic.py
@@ -45,12 +45,6 @@ class TestTimestampArithmetic:
             r"\<-?\d+ \* Days\> and \d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} "
             "will overflow"
         )
-        lmsg = "|".join(
-            [
-                "Python int too large to convert to C (long|int)",
-                "int too big to convert",
-            ]
-        )
         lmsg2 = r"Cannot cast <-?20169940 \* Days> to unit=ns without overflow"
 
         with pytest.raises(OutOfBoundsTimedelta, match=lmsg2):
@@ -68,13 +62,14 @@ class TestTimestampArithmetic:
         stamp = Timestamp("2000/1/1")
         offset_overflow = to_offset("D") * 100**5
 
-        with pytest.raises(OverflowError, match=lmsg):
+        lmsg3 = r"Cannot cast <-?10000000000 \* Days> to unit=ns without overflow"
+        with pytest.raises(OutOfBoundsTimedelta, match=lmsg3):
             stamp + offset_overflow
 
         with pytest.raises(OverflowError, match=msg):
             offset_overflow + stamp
 
-        with pytest.raises(OverflowError, match=lmsg):
+        with pytest.raises(OutOfBoundsTimedelta, match=lmsg3):
             stamp - offset_overflow
 
     def test_overflow_timestamp_raises(self):