```
from datetime import timedelta
from pandas._libs.tslibs.timedeltas import *

td = timedelta(days=4)

%timeit delta_to_nanoseconds(td)
1.74 µs ± 25.2 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)  # <- main
300 ns ± 3.38 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)  # <- PR
265 ns ± 7.28 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)  # <- 1.4.2
```